fix: small fixes

+++ src/main/java/com/deloitte/deloitter/dwitt/repository/DwittRepository.java
@@ -4,9 +4,7 @@ import com.deloitte.deloitter.dwitt.entity.Dwitt;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
-public interface DwittRepository extends JpaRepository<Dwitt, Integer> {
+public interface DwittRepository extends JpaRepository<Dwitt, Long> {
 
 }
